[refactor]: isolate the H2Settings section from ProgramNew if HNQIS option is selected for Program Type Dropdown.
